text-predict-cloudfunction-langchain/requirements.txt
functions-framework==3.*
langchain 
google-cloud-aiplatform
google-cloud-logging